Refactored More Toggle Settings popover
