+++ src/ui/popup/components/header/index.tsx
@@ -1,6 +1,8 @@
 import {html} from 'malevic';
 import SiteToggle from '../site-toggle';
-import {CheckBox, Shortcut, TimeRangePicker, Toggle} from '../../../controls';
+import getTimeIcon from './time-icon';
+import MoreToggleSettings from './more-toggle-settings';
+import {Shortcut, Toggle} from '../../../controls';
 import {getLocalMessage} from '../../../../utils/locales';
 import {ExtWrapper, TabInfo} from '../../../../definitions';
 
@@ -13,33 +15,7 @@ type HeaderProps = ExtWrapper & {
     onMoreToggleSettingsClick: () => void;
 };
 
-function getTimeIcon() {
-    const time = new Date();
-    const hours = time.getHours();
-    const minutes = time.getMinutes();
-
-    const cx = 8;
-    const cy = 8;
-    const lh = 5;
-    const lm = 7;
-    const ah = ((hours > 11 ? hours - 12 : hours) + minutes / 60) / 12 * Math.PI * 2;
-    const am = minutes / 60 * Math.PI * 2;
-    const hx = cx + lh * Math.sin(ah);
-    const hy = cy - lh * Math.cos(ah);
-    const mx = cx + lm * Math.sin(am);
-    const my = cy - lm * Math.cos(am);
-
-    return [
-        `url('`,
-        `data:image/svg+xml;utf-8,`,
-        `<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 16 16">`,
-        `<path fill="none" stroke="white" stroke-width="2" d="M${hx},${hy} L${cx},${cy} L${mx},${my}" />`,
-        `</svg>`,
-        `')`,
-    ].join('');
-}
-
-export function Header({data, actions, tab, onMoreToggleSettingsClick}: HeaderProps) {
+function Header({data, actions, tab, onMoreToggleSettingsClick}: HeaderProps) {
 
     function toggleExtension(enabled) {
         actions.changeSettings({
@@ -106,34 +82,7 @@ export function Header({data, actions, tab, onMoreToggleSettingsClick}: HeaderPr
     );
 }
 
-export function MoreToggleSettings({data, actions, isExpanded, onClose}: ExtWrapper & {isExpanded, onClose}) {
-    return (
-        <div
-            class={{
-                'header__app-toggle__more-settings': true,
-                'header__app-toggle__more-settings--expanded': isExpanded,
-            }}
-        >
-            <div class="header__app-toggle__more-settings__top">
-                <span class="header__app-toggle__more-settings__top__text">{getLocalMessage('time_settings')}</span>
-                <span class="header__app-toggle__more-settings__top__close" role="button" onclick={onClose}>✕</span>
-            </div>
-            <div class="header__app-toggle__more-settings__content">
-                <div class="header__app-toggle__more-settings__line">
-                    <CheckBox
-                        checked={data.settings.automation === 'time'}
-                        onchange={(e) => actions.changeSettings({automation: e.target.checked ? 'time' : ''})}
-                    />
-                    <TimeRangePicker
-                        startTime={data.settings.time.activation}
-                        endTime={data.settings.time.deactivation}
-                        onChange={([start, end]) => actions.changeSettings({time: {activation: start, deactivation: end}})}
-                    />
-                </div>
-                <p class="header__app-toggle__more-settings__description">
-                    {getLocalMessage('set_active_hours')}
-                </p>
-            </div>
-        </div>
-    );
-}
+export {
+    Header,
+    MoreToggleSettings, // TODO: Implement portals to place elements into <body>.
+};
